+++ jax/experimental/mosaic/gpu/utils.py
@@ -1052,6 +1052,23 @@ class BarrierRef:
         1,
     )
 
+  def test_parity(self, parity, orders_tensor_core=False) -> ir.Value:
+    i32 = ir.IntegerType.get_signless(32)
+    parity = arith.extui(i32, parity)
+    wait_complete = nvvm.mbarrier_test_wait(self.get_ptr(), parity)
+    if orders_tensor_core:
+      with when(wait_complete):
+        nvvm.tcgen05_fence(nvvm.Tcgen05FenceKind.AFTER_THREAD_SYNC)
+    return wait_complete
+
+  def test(self, orders_tensor_core: bool = False) -> ir.Value:
+    parities = memref.load(self.phases, [])
+    parity, new_parities = self.update_parities(parities)
+    wait_complete = self.test_parity(parity, orders_tensor_core)
+    with when(wait_complete):
+      memref.store(new_parities, self.phases, [])
+    return wait_complete
+
   def wait_parity(self, parity, orders_tensor_core=False):
     i32 = ir.IntegerType.get_signless(32)
     ticks = arith.constant(i32, 10000000)
@@ -1192,6 +1209,15 @@ class DialectBarrierRef:
 
   def __getitem__(self, offset: ir.Value | int) -> "DialectBarrierRef":
     return DialectBarrierRef(self.barrier_ref[offset], self.orders_tensor_core)
+
+  def test_parity(self, parity, orders_tensor_core=False) -> ir.Value:
+    assert self.orders_tensor_core == orders_tensor_core
+    return self.barrier_ref.test_parity(parity, orders_tensor_core)
+
+  def test(self, orders_tensor_core: bool = False) -> ir.Value:
+    assert self.orders_tensor_core == orders_tensor_core
+    assert self.barrier_ref.phases is not None
+    return self.barrier_ref.test(orders_tensor_core)
 
   def wait_parity(self, parity, orders_tensor_core=False):
     assert self.orders_tensor_core == orders_tensor_core